requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.11.7
python-dotenv==1.0.0
httpx==0.28.1
python-multipart==0.0.6

